fix sdl2 backend, add ncurses and caca backends (not finished)
ncurses backend is fast enough, but will only work with GB.

caca is slow, like, very very slow. it does have the option to spawn its
own window, which i belivie it hw accel the rendering, which is then fullspeed.
however, the font size is too big so the game window is huge. not sure how
to reduce font size yet...

sdl2 backend wasn't updated to support some of the api changes i made, its now
changed to support it.

added some more QOL changes to the cmake, now you can set the video / audio
backend, which the value is then checked if its in supported list of VO.